Shortened excessive dialogue
I edited the dialogue, specifically the lines that show up at the beginning of the game, to be shorter. This is a game that encourages multiple playthroughs, but it is annoying to do so when you are hit by the same worldbuilding monologue at the beginning every time. This should solve the issue of the game not feeling like it has enough choice.